+++ server/models/user.js
@@ -19,23 +19,25 @@ const UserSchema = new mongoose.Schema({
   postsFavourited: [{ type: mongoose.Schema.Types.ObjectId }],
   demo: [],
   profile: {
-    firstName: String,
-    middleName: String,
-    lastName: String,
-    DOB: String,
-    gender: String,
-    contactNo: String,
-    branch: String,
-    currentYear: String,
-    Email: String,
-    degree: String,
-    cgpa: String,
-    joiningYear: String,
-    higherSecondaryMarks: String,
-    higherSecondaryYear: String,
-    secondaryMarks: String,
-    secondaryYear: String,
-
+    type: mongoose.Schema.Types.Mixed,
+    default: {
+      firstName: '',
+      middleName: '',
+      lastName: '',
+      DOB: '',
+      gender: '',
+      contactNo: '',
+      branch: '',
+      currentYear: '',
+      Email: '',
+      degree: '',
+      cgpa: '',
+      joiningYear: '',
+      higherSecondaryMarks: '',
+      higherSecondaryYear: '',
+      secondaryMarks: '',
+      secondaryYear: '',
+    },
   },
 }, options);
 /**
